Refactores game logic
