+++ src/components/app/App.tsx
@@ -1,24 +1,12 @@
-import { MAXIMUM_POINTS, Result, Round, Score } from "components/common";
 import React, { useState } from "react";
+import { Round, Score } from "components/common";
 import { Keypad } from "components/keypad";
 import { Phase } from "./types";
 import { Scoreboard } from "components/scoreboard";
+import { determineResult } from "./logic";
 import dropRight from "lodash.dropright";
 import last from "lodash.last";
 import styles from "./App.module.css";
-
-const determineResult = (bid: Score, winner: Score): Result => {
-  const result = {
-    [winner.team]: winner.points,
-    [winner.team === "us" ? "they" : "us"]: MAXIMUM_POINTS - winner.points
-  };
-
-  if (result[bid.team] < bid.points) {
-    result[bid.team] = -bid.points;
-  }
-
-  return { they: result.they, us: result.us };
-};
 
 export const App: React.FC = () => {
   const [phase, setPhase] = useState<Phase>("bid");
